requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
chess==1.7.0
chess.com==1.7.5
configparser==5.0.2
cssselect2==0.4.1
defusedxml==0.7.1
discord==1.7.3
discord-components==2.1.2
discord-webhook==0.14.0
discord.py==1.7.3
emoji==1.5.0
idna==3.2
lxml==4.9.1
multidict==5.1.0
nekos.py==1.0.3
Pillow==9.0.1
pornhub-api==0.2.0
pornhubapi==0.1.0
psycopg2==2.9.1
pycparser==2.21
pydantic==1.8.2
requests==2.26.0
soupsieve==2.2.1
tinycss2==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
yarl==1.6.3
